Aggiunto convertitore
Permette di convertire il file csv ottenuto in una collezione di classi, tenendo conto di eventuali errori nel file originale.

+++ DemanioFluviale/src/InputOutput/Downloader.java
@@ -1,17 +1,20 @@
 package InputOutput;
+
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 
 import org.apache.commons.io.FileUtils;
-import org.jsoup.Jsoup;
+import org.json.*;
+import org.jsoup.*;
 import org.jsoup.nodes.Document;
-import org.jsoup.select.Elements;
+
+
 
 /**
  * Con questa classe si può ricercare un url e scaricarne il contenuto
- * @author 
+ * @author Real Estate Group
  * 
  */
 
@@ -29,46 +32,55 @@ public class Downloader {
 	 */
 	
 	public Downloader(URL url, String ext) {
-		setUrl(url); setArg(ext);
+		setUrl(url); setExt(ext);
 	}
 	
 	/**
 	 * partendo dal file descrittivo del data-set ricerca l'url assoluto del file da scaricare
 	 * @return false: nessun file da scaricare trovato
 	 * @throws IOException
+	 * @throws JSONException 
 	 * 
 	 */
-	public boolean RicercaFile() throws IOException {
+	public boolean RicercaFile() throws IOException, JSONException {
 	    
-		String links = null;
-		Document doc = Jsoup.connect(getUrl().toString()).get();
-		Elements link = doc.select("a[href]");
+		Document doc = Jsoup.connect(getUrl().toString()).ignoreContentType(true).get();
+		JSONObject obj = new JSONObject(doc.body().text()).getJSONObject("result");
+		JSONArray array = obj.getJSONArray("resources");
 		
-		for(int i = 0; i < link.size(); i++) {
-			links = link.get(i).absUrl("href");
-			if(links.equalsIgnoreCase(".csv")) {setFile_url(links); return true;}
+		for(int i = 0; i < array.length(); i++) {
+			String temp_url = array.getJSONObject(i).getString("url");
+			if(temp_url.contains(getExt())) { setFile_url(array.getJSONObject(i).getString("url")); return true;}
 		}
 		
 		return false;
 	}
 	
-	public void url2file(String filename) throws MalformedURLException, IOException {
+	/**
+	 * converte, se necessario, l'url immesso in un file csv
+	 * @param filename
+	 * @throws MalformedURLException
+	 * @throws IOException
+	 * @throws JSONException
+	 */
+	public void url2file(String filename) throws MalformedURLException, IOException, JSONException {
+		
+		if(new File(filename).exists()) {System.out.println("Non serve effettuare il download, file gia' presente\n"); return;};
 		
 		if(RicercaFile()) {
 			FileUtils.copyURLToFile(new URL(getFile_url()), new File(filename));
 			System.out.println("Download effettuato da: " + getFile_url() + ", file CSV generato\n");
 		}
-		
 		else System.out.println("Download non effettuato\n");
 	}
 	
 	
-	
-	public String getArg() {
+	/* GETTER E SETTER VARI */
+	public String getExt() {
 		return ext;
 	}
 	
-	public void setArg(String ext) {
+	public void setExt(String ext) {
 		this.ext = ext;
 	}
 	
@@ -87,4 +99,5 @@ public class Downloader {
 	public void setFile_url(String file_url) {
 		this.file_url = file_url;
 	}
+	/**************************/
 }
